+++ src/Models/States/DevicePropertiesManager.php
@@ -18,6 +18,7 @@ namespace FastyBird\MiniServer\Models\States;
 use FastyBird\DevicesModule\Entities as DevicesModuleEntities;
 use FastyBird\DevicesModule\Models as DevicesModuleModels;
 use FastyBird\DevicesModule\States as DevicesModuleStates;
+use FastyBird\MiniServer\Exceptions;
 use FastyBird\MiniServer\States;
 use FastyBird\RedisDbStoragePlugin\Models as RedisDbStoragePluginModels;
 use Nette;
@@ -63,6 +64,10 @@ class DevicePropertiesManager implements DevicesModuleModels\States\IDevicePrope
 		DevicesModuleStates\IDeviceProperty $state,
 		Utils\ArrayHash $values
 	): States\IDeviceProperty {
+		if (!$state instanceof States\IDeviceProperty) {
+			throw new Exceptions\InvalidArgumentException('Provided state entity is not valid instance');
+		}
+
 		/** @var States\IDeviceProperty $updatedState */
 		$updatedState = $this->statesManager->update(
 			$state,
@@ -76,6 +81,10 @@ class DevicePropertiesManager implements DevicesModuleModels\States\IDevicePrope
 		DevicesModuleEntities\Devices\Properties\IProperty $property,
 		DevicesModuleStates\IDeviceProperty $state
 	): bool {
+		if (!$state instanceof States\IDeviceProperty) {
+			return false;
+		}
+
 		return $this->statesManager->delete($state);
 	}
